==> R/06_Figure_1.R <==
source(here::here("R", "00_setup.R"))

# A map of Sierra Leone in Africa -----------------------------------------

world <- ne_countries(scale = "medium", returnclass = "sf")

africa <- world %>%
  filter(str_detect(continent, "Africa")) %>%
  select(admin) %>%
  mutate(fill = case_when(str_detect(admin, "Sierra Leone") ~ "black",
                            TRUE ~ "grey")) %>%
  ms_filter_islands(min_area = 1e10)

sl_bbox <- africa %>%
  filter(str_detect(admin, "Sierra Leone")) %>%
  st_bbox() %>%
  st_as_sfc()

africa_map <- ggplot() +
  geom_sf(data = africa, aes(fill = fill)) +
  geom_sf(data = sl_bbox, fill = NA, colour = "black", size = 4) +
  scale_fill_manual(values = c("grey", "white")) +
  guides(fill = "none") +
  theme_void()

sle_sf <- geodata::gadm(country = "SLE", level = 2, path = here("data", "geodata")) %>%
  st_as_sf()

fig_1_palette <- c(village_palette, "#FFFFFF")
names(fig_1_palette) <- c(names(village_palette)[1:4], "poi")

poi <- tibble(name = c("Kenema", "Baiama", "Lalehun", "Lambayama", "Seilama"),
              cat = c("poi", "Baiama", "Lalehun", "Lambayama", "Seilama"),
              lat = c(7.876161956810467, 7.837529372181356, 8.197392257077409, 7.850593096948891, 8.12230048178563),
              lon = c(-11.190811585001954, -11.268407665149846, -11.08032958100431, -11.196939025872055, -11.1935976318381)) %>%
  st_as_sf(coords = c("lon", "lat"), crs = default_CRS)


# Focus map on Kenema
kenema_map <- get_googlemap(st_coordinates(poi %>% filter(name == "Kenema"))[1, ], zoom = 9, maptype = "satellite", scale = 2) %>%
  rast() %>%
  mask(vect(sle_sf %>%
              filter(str_detect(NAME_2, "Kenema"))))

zoom_kenema <- sle_sf %>%
  filter(str_detect(NAME_2, "Kenema")) %>% st_bbox()

sle_map <- ggplot() +
  geom_sf(data = sle_sf) +
  geom_sf(data = sle_sf %>%
            filter(str_detect(NAME_2, "Kenema")), fill = "orange") +
  geom_sf(data = zoom_kenema %>%
            st_as_sfc(), fill = NA, colour = "black", size = 4)  +
  guides(fill = "none") +
  theme_void()

study_map <- ggplot() + 
  geom_sf(data = sle_sf, fill = "grey", colour = "black") +
  geom_spatraster_rgb(data = kenema_map) +
  geom_sf(data = sle_sf %>%
            filter(str_detect(NAME_2, "Kenema")), fill = NA, colour = "white") +
  geom_sf(data = poi, size = 0.8) +
  ggrepel::geom_label_repel(data = poi, aes(label = name, geometry = geometry, fill = cat), stat = "sf_coordinates", min.segment.length = 0) +
  labs(x = element_blank(),
       y = element_blank()) +
  scale_fill_manual(values = fig_1_palette) +
  theme_bw() +
  ggspatial::annotation_north_arrow(style = north_arrow_minimal()) +
  ggspatial::annotation_scale(location = "br") +
  coord_sf(xlim = c(zoom_kenema[1], zoom_kenema[3]), ylim = c(zoom_kenema[2], zoom_kenema[4])) +
  guides(fill = "none") +
  labs(title = "A)")

sl_inset_map <- ggdraw() +
  draw_plot(study_map) +
  draw_plot(africa_map, x = 0.7, y = 0.7, width = 0.3, height = 0.3) +
  draw_plot(sle_map, x = 0.7, y = 0.4, width = 0.3, height = 0.3)

ggsave2(plot = sl_inset_map, filename = here("output", "Figure_1a.png"), dpi = 300, width = 8, height = 6)

# Trap timeline -----------------------------------------------------------
trap_data <- read_rds(here("data", "input", "combined_data.rds"))
trap_data <- trap_data$trap_data

timeline <- trap_data %>%
  tibble() %>%
  select(-geometry) %>%
  filter(village != "bambawo") %>%
  group_by(village, visit, grid_number, trap_number) %>%
  summarise(tn = n(),
            date_set = min(date_set)) %>%
  group_by(visit, village) %>%
  summarise(tn = sum(tn),
            date_set = min(date_set)) %>%
  ungroup() %>%
  mutate(village = str_to_title(village),
         date_set = case_when(date_set == as.Date("2023-02-05") ~ as.Date("2023-02-08"),
                              TRUE ~ date_set))

rain_season <- tibble(season = "rainy",
                      date_start = c(as_date("2021-05-01"), as_date("2022-05-01"), as_date("2023-05-01")),
                      date_end = c(as_date("2021-11-01"), as_date("2022-11-01"), as_date("2023-11-01")))

timeline_plot <- ggplot(timeline) +
  geom_rect(aes(xmin = date_set, xmax = date_set + 4, ymin = 0, ymax = tn, fill = village)) +
  geom_rect(data = rain_season, aes(xmin = date_start, xmax = date_end, ymin = 0, ymax = Inf), fill = "orange", alpha = 0.1) +
  coord_cartesian(xlim = c(as.Date("2020-11-01"), as.Date("2023-05-01"))) +
  scale_fill_manual(values = fig_1_palette) +
  theme_bw() +
  labs(x = "Visit date",
       y = "Trap nights",
       fill = "Village") +
  labs(title = "B)")

save_plot(plot = timeline_plot, filename = here("output", "Figure_1b.png"), base_height = 3, base_width = 8)

save_plot(plot = plot_grid(plotlist = list(sl_inset_map,
                                    timeline_plot),
                    ncol = 1,
                    rel_heights = c(8, 2)),
          filename = here("output", "Figure_1_combined.png"),
          base_height = 10,
          base_width = 8)


### Sort out the remaining locations for the supplementary

# Trap locations ----------------------------------------------------------

fig_1_df <- read_rds(here("data", "observed_data", "fig_1_df.rds"))

bbox <- list()

bbox$baiama <- st_bbox(st_buffer(fig_1_df[[1]], dist = 100))
bbox$lalehun <- st_bbox(st_buffer(fig_1_df[[2]], dist = 100))
bbox$lambayama <- st_bbox(st_buffer(fig_1_df[[3]], dist = 100))
bbox$seilama <- st_bbox(st_buffer(fig_1_df[[4]], dist = 100))

bg <- list()

bg$baiama <- get_googlemap(c(-11.26, 7.832),
                           zoom = 15,
                           scale = 2,
                           maptype = "satellite") %>%
  rast() %>%
  mask(vect(st_as_sfc(bbox$baiama) %>%
              st_transform(crs = default_CRS)))

bg$lalehun <- get_googlemap(poi %>%
                              filter(str_detect(name, "Lalehun")) %>%
                              st_coordinates(.),
                            zoom = 15,
                            scale = 2,
                            maptype = "satellite") %>%
  rast() %>%
  mask(vect(st_as_sfc(bbox$lalehun) %>%
              st_transform(crs = default_CRS)))

bg$lambayama <- get_googlemap(poi %>%
                                filter(str_detect(name, "Lambayama")) %>%
                                st_coordinates(.),
                              zoom = 16,
                              scale = 2,
                              maptype = "satellite") %>%
  rast() %>%
  mask(vect(st_as_sfc(bbox$lambayama) %>%
              st_transform(crs = default_CRS)))

bg$seilama <- get_googlemap(poi %>%
                              filter(str_detect(name, "Seilama")) %>%
                              st_coordinates(.),
                            zoom = 16,
                            scale = 2,
                            maptype = "satellite") %>%
  rast() %>%
  mask(vect(st_as_sfc(bbox$seilama) %>%
              st_transform(crs = default_CRS)))

grids_plot <- list()

breaks <- list(baiama = list(x = c(-11.265, -11.255),
                             y = c(7.838, 7.83, 7.824)),
               lalehun = list(x = c(-11.081, -11.079),
                              y = c(8.194, 8.196, 8.199)),
               lambayama = list(x = c(-11.198, -11.195, -11.192),
                                y = c(7.8515, 7.85, 7.8485)),
               seilama = list(x = c(-11.198, -11.195),
                              y = c(8.124, 8.122)))

for(i in 1:length(fig_1_df))  {
  
  grids_plot[[i]] <- ggplot() + 
    geom_spatraster_rgb(data = bg[[i]]) +
    geom_sf(data = fig_1_df[[i]] %>%
              mutate(landuse = factor(str_to_title(landuse), levels = c("Forest", "Agriculture", "Village"))),
            aes(fill = tn, colour = tn)) +
    scale_colour_viridis_c(limits = c(0, 100), direction = -1) +
    scale_fill_viridis_c(limits = c(0, 100), direction = -1) +
    facet_wrap(~ landuse) +
    labs(fill = "Number Trap-Nights",
         title = str_to_title(unique(fig_1_df[[i]]$village))) +
    coord_sf(xlim = ext(vect(st_buffer(fig_1_df[[i]], dist = 100) %>% st_transform(crs = default_CRS)))[1:2],
             ylim = ext(vect(st_buffer(fig_1_df[[i]], dist = 100) %>% st_transform(crs = default_CRS)))[3:4],
             expand = FALSE) +
    scale_x_continuous(breaks = breaks[[i]]$x) +
    scale_y_continuous(breaks = breaks[[i]]$y) +
    theme_bw() +
    annotation_scale()
  
}

save_plot(plot = grids_plot[[1]] +
            labs(title = "A)") +
            guides(fill = "none"), filename = here("output", "Supplementary_Figure_1a.png"), base_height = 3, base_width = 7)
save_plot(plot = grids_plot[[2]] +
            labs(title = "B)") +
            guides(fill = "none"), filename = here("output", "Supplementary_Figure_1b.png"), base_height = 4, base_width = 6)
save_plot(plot = grids_plot[[3]] +
            labs(title = "C)") +
            guides(fill = "none"), filename = here("output", "Supplementary_Figure_1c.png"), base_height = 3, base_width = 7)
save_plot(plot = grids_plot[[4]] +
            labs(title = "D)") +
            guides(fill = "none"), filename = here("output", "Supplementary_Figure_1d.png"), base_height = 3, base_width = 7)
